Cleanup of some stuff after due time
I just wanted to work on this to clean it up, since there were some minor mistakes bugging me, and this is simple stuff that I would have gotten with a little more time testing. I still haven't worked on the keystroke interruption requirement, and will leave that for later if I ever feel up for adding it; I also would like to maybe add some graphics interaction to display the error window, but that would be down the road.